# services/authentication-api/swagger-2.0.yaml
swagger: '2.0'
info:
  title: Authentication API
  description: Swagger 2.0 specification for the authentication-api service
  version: '1.0.0'
host: localhost:3000
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

definitions:
  SignInRequest:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
      password:
        type: string

  SignUpRequest:
    type: object
    required:
      - email
      - password
      - userName
    properties:
      email:
        type: string
        format: email
      password:
        type: string
      userName:
        type: string

  TokenRequest:
    type: object
    required:
      - token
    properties:
      token:
        type: string

  EmailRequest:
    type: object
    required:
      - email
    properties:
      email:
        type: string
        format: email

  VerifyRecoveryCodeRequest:
    type: object
    required:
      - email
      - code
    properties:
      email:
        type: string
        format: email
      code:
        type: string

  ResetPasswordRequest:
    type: object
    required:
      - token
      - newPassword
    properties:
      token:
        type: string
      newPassword:
        type: string

  CreateUserRequest:
    type: object
    required:
      - name
      - email
      - password
      - recoveryToken
      - recoveryTokenExpiration
    properties:
      name:
        type: string
      email:
        type: string
        format: email
      password:
        type: string
      recoveryToken:
        type: string
      recoveryTokenExpiration:
        type: string

  User:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
        format: email
      recoveryToken:
        type: string
      recoveryTokenExpiration:
        type: string
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time

  DeleteResponse:
    type: object
    properties:
      message:
        type: string

paths:
  /:
    get:
      tags:
        - Health
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Successful response with greeting text.
          schema:
            type: string
          examples:
            application/json: Hello World!

  /auth/signin:
    post:
      tags:
        - Auth
      summary: Sign in user
      operationId: signIn
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignInRequest'
      responses:
        '200':
          description: Successful sign-in response.
          schema:
            type: object

  /auth/signup:
    post:
      tags:
        - Auth
      summary: Sign up new user
      operationId: signUp
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignUpRequest'
      responses:
        '201':
          description: User created successfully.
          schema:
            type: object

  /auth/signout:
    post:
      tags:
        - Auth
      summary: Sign out user
      operationId: signOut
      responses:
        '200':
          description: Successful sign-out response.
          schema:
            type: object

  /auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TokenRequest'
      responses:
        '200':
          description: Token refreshed successfully.
          schema:
            type: object

  /auth/send-recovery-code:
    post:
      tags:
        - Auth
      summary: Send password recovery code to email
      operationId: sendRecoveryCode
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/EmailRequest'
      responses:
        '200':
          description: Recovery code sent successfully.
          schema:
            type: object

  /auth/verify-recovery-code:
    post:
      tags:
        - Auth
      summary: Verify password recovery code
      operationId: verifyRecoveryCode
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/VerifyRecoveryCodeRequest'
      responses:
        '200':
          description: Recovery code verified successfully.
          schema:
            type: object

  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset user password
      operationId: resetPassword
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully.
          schema:
            type: object

  /auth/google:
    post:
      tags:
        - Auth
      summary: Sign in with Google
      operationId: googleSignIn
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TokenRequest'
      responses:
        '200':
          description: Successful Google sign-in response.
          schema:
            type: object

  /users:
    get:
      tags:
        - Users
      summary: Get all users
      operationId: findAllUsers
      responses:
        '200':
          description: List of users.
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateUserRequest'
      responses:
        '201':
          description: User created successfully.
          schema:
            $ref: '#/definitions/User'

  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      operationId: findOneUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        '200':
          description: User found.
          schema:
            $ref: '#/definitions/User'
    post:
      tags:
        - Users
      summary: Update user by ID
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateUserRequest'
      responses:
        '200':
          description: User updated successfully.
          schema:
            $ref: '#/definitions/User'

  /users/{id}/patch:
    post:
      tags:
        - Users
      summary: Partially update user by ID
      operationId: patchUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateUserRequest'
      responses:
        '200':
          description: User patched successfully.
          schema:
            $ref: '#/definitions/User'

  /users/{id}/delete:
    post:
      tags:
        - Users
      summary: Delete user by ID
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        '200':
          description: User deleted successfully.
          schema:
            $ref: '#/definitions/DeleteResponse'